pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kon"]

[tool.uv]
package = true

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true

[project]
name = "kon-coding-agent"
version = "0.1.0"
description = "Minimal coding agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.3",
    "anthropic>=0.79.0",
    "openai>=2.21.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "rich>=14.3.2",
    "textual>=8.0.0",
]

[dependency-groups]
dev = [
 "pyright>=1.1.408",
 "pytest>=9.0.2",
 "pytest-asyncio>=1.3.0",
]


[project.scripts]
kon = "kon.ui.app:main"

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
